Fix tile blending in BFME by using correct bitsize for Blends, ThreeWayBlends, and CliffTextures

+++ src/OpenSage.Game/Content/MapLoader.cs
@@ -601,7 +601,7 @@ namespace OpenSage.Content
         private static BlendData GetBlendData(
             MapFile mapFile,
             int x, int y,
-            ushort blendIndex,
+            uint blendIndex,
             byte baseTextureIndex)
         {
             if (blendIndex > 0)
